Adopt Axel's solution to the unmount/rmdir problem (only remove from rootfs) in DriveSetup also.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@26611 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/drivesetup/MainWindow.cpp
@@ -633,7 +633,8 @@ MainWindow::_Unmount(BDiskDevice* disk, partition_id selectedPartition)
 			_DisplayPartitionError("Could not unmount partition %s.",
 				partition, ret);
 		} else {
-			rmdir(path.Path());
+			if (dev_for_path(path.Path()) == dev_for_path("/"))
+				rmdir(path.Path());
 			// successful unmount, adapt to the changes
 			_ScanDrives();
 		}
